refactor(packages): Split markdown inputter and package
